+++ src/jndev/pseudo3d/sprite/AnimatedSprite.java
@@ -17,7 +17,7 @@ public class AnimatedSprite extends Sprite {
     /**
      * current frame number
      */
-    private double currentFrame;
+    private float currentFrame;
     
     /**
      * all images of the animated sprite
@@ -27,7 +27,7 @@ public class AnimatedSprite extends Sprite {
     /**
      * time between frames
      */
-    private final double frameStep;
+    private final float frameStep;
     
     /**
      * create a new animated sprite with a list of images and frame rate
@@ -35,9 +35,9 @@ public class AnimatedSprite extends Sprite {
      * @param images    all images of the animated sprite
      * @param frameRate frames per second of the sprite
      */
-    public AnimatedSprite(ArrayList<BufferedImage> images, double frameRate) {
+    public AnimatedSprite(ArrayList<BufferedImage> images, float frameRate) {
         super(images.get(0));
-        this.frameStep = frameRate / 1000.0;
+        this.frameStep = frameRate / 1000.0f;
         currentFrame = 0;
         this.images = images;
     }
@@ -58,7 +58,7 @@ public class AnimatedSprite extends Sprite {
      * set the current frame to the next available frame based on elapsed time
      */
     public void update() {
-        double renderStep = Pseudo3D.getInstance().getGameLoop().getRenderFrequency() / 1000.0;
+        float renderStep = Pseudo3D.getInstance().getGameLoop().getRenderFrequency() / 1000.0f;
         currentFrame = currentFrame + (frameStep / renderStep) <
                 images.size() ? currentFrame + (frameStep / renderStep) : 0;
         image = images.get((int) FastMath.floor(currentFrame));
@@ -76,8 +76,8 @@ public class AnimatedSprite extends Sprite {
         if (o == null || getClass() != o.getClass()) return false;
         if (!super.equals(o)) return false;
         AnimatedSprite that = (AnimatedSprite) o;
-        return Double.compare(that.currentFrame, currentFrame) == 0 &&
-                Double.compare(that.frameStep, frameStep) == 0 &&
+        return Float.compare(that.currentFrame, currentFrame) == 0 &&
+                Float.compare(that.frameStep, frameStep) == 0 &&
                 Objects.equals(images, that.images) &&
                 super.equals(that);
     }
